+++ src/components/common/DeviceInfo.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script setup lang="ts">
-import type { DeviceStats } from '@/schemas/apiTypes'
+import type { DeviceStats } from '@/schemas/apiSchema'
 import { formatSize } from '@/utils/formatUtil'
 
 const props = defineProps<{
