+++ src/components/common/SystemStatsPanel.vue
@@ -36,7 +36,7 @@ import TabView from 'primevue/tabview'
 import { computed } from 'vue'
 
 import DeviceInfo from '@/components/common/DeviceInfo.vue'
-import type { SystemStats } from '@/schemas/apiTypes'
+import type { SystemStats } from '@/schemas/apiSchema'
 import { formatSize } from '@/utils/formatUtil'
 
 const props = defineProps<{
