+++ src/components/sidebar/tabs/nodeLibrary/NodeTreeFolder.vue
@@ -11,7 +11,7 @@
 import { Ref, computed, inject, onMounted, onUnmounted, ref, watch } from 'vue'
 
 import TreeExplorerTreeNode from '@/components/common/TreeExplorerTreeNode.vue'
-import type { BookmarkCustomization } from '@/schemas/apiTypes'
+import type { BookmarkCustomization } from '@/schemas/apiSchema'
 import { useNodeBookmarkStore } from '@/stores/nodeBookmarkStore'
 import { ComfyNodeDefImpl } from '@/stores/nodeDefStore'
 import { RenderedTreeExplorerNode } from '@/types/treeExplorerTypes'
