+++ src/composables/widgets/useMarkdownWidget.ts
@@ -8,7 +8,7 @@ import TiptapTableRow from '@tiptap/extension-table-row'
 import TiptapStarterKit from '@tiptap/starter-kit'
 import { Markdown as TiptapMarkdown } from 'tiptap-markdown'
 
-import type { InputSpec } from '@/schemas/apiTypes'
+import type { InputSpec } from '@/schemas/apiSchema'
 import type { ComfyWidgetConstructor } from '@/scripts/widgets'
 import type { ComfyApp } from '@/types'
