+++ src/extensions/core/uploadAudio.ts
@@ -6,7 +6,7 @@ import { useNodeDragAndDrop } from '@/composables/node/useNodeDragAndDrop'
 import { useNodeFileInput } from '@/composables/node/useNodeFileInput'
 import { useNodePaste } from '@/composables/node/useNodePaste'
 import { t } from '@/i18n'
-import { ComfyNodeDef } from '@/schemas/apiTypes'
+import { ComfyNodeDef } from '@/schemas/apiSchema'
 import type { DOMWidget } from '@/scripts/domWidget'
 import { useToastStore } from '@/stores/toastStore'
