+++ src/scripts/ui/settings.ts
@@ -1,4 +1,4 @@
-import type { Settings } from '@/schemas/apiTypes'
+import type { Settings } from '@/schemas/apiSchema'
 import type { ComfyApp } from '@/scripts/app'
 import { useSettingStore } from '@/stores/settingStore'
 import { useToastStore } from '@/stores/toastStore'
