+++ src/stores/userStore.ts
@@ -1,7 +1,7 @@
 import { defineStore } from 'pinia'
 import { computed, ref, watchEffect } from 'vue'
 
-import type { User as UserConfig } from '@/schemas/apiTypes'
+import type { User as UserConfig } from '@/schemas/apiSchema'
 import { api } from '@/scripts/api'
 
 export interface User {
