+++ src/types/litegraph-augmentation.d.ts
@@ -1,7 +1,7 @@
 import '@comfyorg/litegraph'
 import type { LLink, Size } from '@comfyorg/litegraph'
 
-import type { ComfyNodeDef } from '@/schemas/apiTypes'
+import type { ComfyNodeDef } from '@/schemas/apiSchema'
 import type { DOMWidget, DOMWidgetOptions } from '@/scripts/domWidget'
 
 import type { NodeId } from '../schemas/comfyWorkflow'
